+++ src/components/Tokens/TokenTable/TokenRow.tsx
@@ -8,7 +8,6 @@ import { getChainInfo } from 'constants/chainInfo'
 import { SparklineMap, TopToken } from 'graphql/data/TopTokens'
 import { CHAIN_NAME_TO_CHAIN_ID, getTokenDetailsURL } from 'graphql/data/util'
 import { useAtomValue } from 'jotai/utils'
-import { useTokenLogoURI } from 'pages/TokenDetails'
 import { ForwardedRef, forwardRef } from 'react'
 import { CSSProperties, ReactNode } from 'react'
 import { ArrowDown, ArrowUp } from 'react-feather'
@@ -32,6 +31,7 @@ import {
   TokenSortMethod,
   useSetSortMethod,
 } from '../state'
+import { useTokenLogoURI } from '../TokenDetails'
 import InfoTip from '../TokenDetails/InfoTip'
 import { formatDelta, getDeltaArrow } from '../TokenDetails/PriceChart'
