+++ src/components/Tokens/state.ts
@@ -11,7 +11,7 @@ export enum TokenSortMethod {
 }
 
 export const filterStringAtom = atomWithReset<string>('')
-export const filterTimeAtom = atom<TimePeriod>(TimePeriod.DAY)
+export const filterTimeAtom = atom<TimePeriod>(TimePeriod.YEAR)
 export const sortMethodAtom = atom<TokenSortMethod>(TokenSortMethod.VOLUME)
 export const sortAscendingAtom = atom<boolean>(false)
